#208 add lazy expansion for unabstracteed state
